==> retro-analysis/01a-nowind-loop.R <==
### created: 11/11/2022
### last updated: 12/08/2022

#### 01a - STRATIFIED CALCULATIONS: WITH WIND AREAS INCLUDED ####

###################
#### OBJECTIVE ####
###################
# calculate stratified means for each species, strata, and year combination based on the historical time series data. The analysis contained herein calculates the stratified means for the full time series without any change to represent the base/control scenario. No tows/observations have been subtracted from this dataset 

# awaiting new full raw dataset from Catherine to run in here(...); when received run and check code. Derived data from tidy-data script will be called to run in the loops. 
####################


#### LOAD PACKAGES ####
library(stringr)
library(sf)
library(patchwork)
library(here)
suppressPackageStartupMessages(library(tidyverse))


#### LOAD DATA ####
# dataset created from `02-complete-dataset.R` here("tidy-data"). Contains complete observations for each species and unique tow. 
data <- readRDS(here("data", "rds", "merged_data_complete.rds")) %>% mutate(EXPCATCHWT = ifelse(is.na(EXPCATCHWT), 0, EXPCATCHWT))

# species dataframe for adding to final dataset 
species <- data %>% 
  select(SVSPP, COMNAME, SCINAME) %>%
  unique()
### save the data 
saveRDS(species, here("data", "rds", "species.rds"))

# load and manipulate the bottom trawl survey strata shapefile
strata <- sf::st_read(dsn = here("gis", "temp", "BTS_Strata", "BTS_Strata.shp")) %>% #read in data
  dplyr::select(Strata_Num, Area_SqNm) %>% # select variables to be used in calculations below
  sf::st_set_geometry(NULL) %>% # remove the coordinates; changes the sf to a df
  unique() %>% # identify unique stratum only 
  rename(STRATUM = Strata_Num) %>% # rename Strat_Num to STRATUM for further analyses below
  mutate(RelWt = Area_SqNm / sum(Area_SqNm)) # calculate the relative weight of each stratum based on its proportion of the total survey footprint area; to be used in later calculations.

### save the data 
saveRDS(strata, here("data", "rds", "strata.rds"))

# calculate total survey area for use in future calculations  
BTSArea <- as.integer(sum(strata$Area_SqNm))

# statistical management areas 
#geounits <- readRDS(here("data", "rds", "geounits.rds"))

#### CALCULATE INDIVIDUAL MEANS AND VARIANCES ####
# calculate individual means and variances for each combinations of species, year, and stratum
nowind_means <- data %>%  
  mutate(code = str_c(STRATUM, CRUISE6, STATION)) %>% 
  left_join(geounits, by = "code") %>% 
  rename(STRATUM = STRATUM.x, 
         STATION = STATION.x) %>% 
  group_by(STRATUM, EST_YEAR, SVSPP, SEASON) %>% #, GEO_AREA) %>% 
  summarise(towct = length(unique(STATION)), # calculate unique tows
            mu = sum(EXPCATCHWT)/towct, # find the average biomass based on unique tows rather than observations to avoid potential duplication 
            var = ifelse(towct == 1, 0, # if the tow count equals 1, then variance about the mean should be 0
                         sum((EXPCATCHWT - mu)^2)/(towct - 1))) %>% # if tow count does not equal 1, then find the variance of biomass
  left_join(strata, by = "STRATUM") %>% # add each stratum area and relative weight to the dataset based on STRATUM number
  mutate(wt_mu = Area_SqNm * mu, # part one of the stratified mean formula
         wt_var = ((((RelWt)^2) * var) / towct) * (1 - (towct / Area_SqNm))) # part one of the stratified variance formula


### save the data 
saveRDS(nowind_means, here("data", "rds", "indiv-mu_included.rds"))



#### COMPLETE STRATIFIED MEAN AND VARIANCE CALCULATIONS ####
# calculate stratified means and variances for each combinations of species and year based on individual stratum means and variances 
nowind_stratmu <- nowind_means %>% 
  group_by(SVSPP, EST_YEAR, SEASON) %>% #, GEO_AREA) %>% 
  summarise(stratmu = (sum(wt_mu)) / BTSArea, # part two of the stratified mean formula
            stratvar = sum(wt_var)) %>% # part two of the stratified variance formula
  mutate(TYPE = paste("With Wind Included")) # paste identifying information of means and variances for joining and plotting in later scripts


### save the data 
saveRDS(nowind_stratmu, here("data", "rds", "strat-mu_included.rds"))
